自动阈值模式下实现控制设备自动下发控制功能完成,采集设备实时在线状态展示待完善

+++ src/main/java/yuan/xu/intelligence_agriculture/dto/SensorData.java
@@ -2,6 +2,8 @@ package yuan.xu.intelligence_agriculture.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SensorData {
     /**
@@ -18,7 +20,7 @@ public class SensorData {
     /**
      * 设备数据
      */
-    private Integer data;
+    private BigDecimal data;
 
 }
